fix(tui): scroll the detail pane past line 65,535
The detail scroll offset was a u16 and the panels scrolled through Paragraph's u16 offset, so no row past line 65,535 could be shown. The offset is now a usize end to end: the coefficient views build only the rows in view (no placeholder lines for the rest), the rows left out stand in for the scroll, and what remains is dropped from the front rather than scrolled by a u16. The summary and raw-text views take the wider offset the same way.

Dropping the placeholders also takes the per-frame cost of the modified views off the model size. render bench, criterion vs the pre-windowing baseline:
  diff_objective_modified_200k   1.94 ms -> 0.51 ms  (-73.9%, p < 0.05)
  diff_constraint_modified_200k  1.58 ms -> 0.10 ms  (-93.7%, p < 0.05)

+++ tui/src/ui.rs
@@ -152,7 +152,7 @@ pub fn draw(frame: &mut Frame, app: &mut App) {
     // same position feedback without needing focus.
     let detail_inner_height = detail_area.height.saturating_sub(2) as usize;
     if app.layout.detail_content_lines > detail_inner_height {
-        let mut scrollbar_state = ScrollbarState::new(app.layout.detail_content_lines).position(app.detail_scroll as usize);
+        let mut scrollbar_state = ScrollbarState::new(app.layout.detail_content_lines).position(app.detail_scroll);
         crate::widgets::render_panel_scrollbar(frame, detail_area, &mut scrollbar_state);
     }
